Small corruption change

+++ src/main/java/tragicneko/tragicmc/entity/mob/EntityPlague.java
@@ -92,7 +92,7 @@ public class EntityPlague extends TragicMob {
 
 		if (this.worldObj.isRemote)
 		{
-			for (int l = 0; l < 2; ++l)
+			for (byte l = 0; l < 2; ++l)
 			{
 				this.worldObj.spawnParticle(EnumParticleTypes.SPELL_WITCH,
 						this.posX + (this.rand.nextDouble() - rand.nextDouble()) * this.width * 1.5D,
